relocated db file to common directory
